requirements/common.txt
Django==3.0.5
pylint==2.4.4
psycopg2-binary==2.8.5
autopep8==1.5.1
djangorestframework==3.11.0